Move custom field builder to its own file
